# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Primeiro Prontuário API
  version: 1.0.0
  description: |
    Contrato completo do MVP acadêmico para um único médico.
    Use somente dados fictícios ou anonimizados.
  license:
    name: Não licenciado - uso acadêmico demonstrativo
    identifier: LicenseRef-Proprietary
servers:
  - url: https://prontuario.local
    description: Cluster Kubernetes local com TLS confiado apenas na máquina de demonstração
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/appointments:
    get:
      tags: [Agenda]
      operationId: searchAppointments
      summary: Consultar agenda
      parameters:
        - $ref: "#/components/parameters/RequiredFrom"
        - $ref: "#/components/parameters/RequiredTo"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AppointmentStatus"
        - name: patientId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Página de agendamentos em ordem de início e id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Agenda]
      operationId: createAppointment
      summary: Agendar paciente
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Agendamento criado
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/appointments/reminders:
    get:
      tags: [Agenda]
      operationId: listAppointmentReminders
      summary: Listar lembretes internos
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Agendamentos na janela de lembrete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/appointments/{appointmentId}:
    parameters:
      - $ref: "#/components/parameters/AppointmentId"
    get:
      tags: [Agenda]
      operationId: getAppointment
      summary: Consultar agendamento
      responses:
        "200":
          description: Agendamento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/appointments/{appointmentId}/schedule:
    parameters:
      - $ref: "#/components/parameters/AppointmentId"
    put:
      tags: [Agenda]
      operationId: rescheduleAppointment
      summary: Reagendar compromisso
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleAppointmentRequest"
      responses:
        "200":
          description: Agendamento reagendado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/appointments/{appointmentId}/status:
    parameters:
      - $ref: "#/components/parameters/AppointmentId"
    patch:
      tags: [Agenda]
      operationId: changeAppointmentStatus
      summary: Transicionar estado do agendamento
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeAppointmentStatusRequest"
      responses:
        "200":
          description: Agendamento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/schedule-blocks:
    get:
      tags: [Bloqueios]
      operationId: searchScheduleBlocks
      summary: Listar bloqueios futuros no período
      parameters:
        - $ref: "#/components/parameters/RequiredFrom"
        - $ref: "#/components/parameters/RequiredTo"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Página solicitada como array ordenado, sem envelope
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleBlock"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [Bloqueios]
      operationId: createScheduleBlock
      summary: Criar bloqueio futuro
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBlockRequest"
      responses:
        "201":
          description: Bloqueio criado
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleBlock"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/schedule-blocks/{blockId}:
    parameters:
      - $ref: "#/components/parameters/BlockId"
    delete:
      tags: [Bloqueios]
      operationId: removeScheduleBlock
      summary: Remover bloqueio futuro
      security:
        - sessionCookie: []
          csrfHeader: []
      responses:
        "204":
          description: Bloqueio removido
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Sessão server-side; HttpOnly, SameSite=Lax e Secure em HTTPS
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        Token devolvido por GET /api/v1/auth/csrf e associado ao cookie
        XSRF-TOKEN do mesmo cliente. Exigido em mutações autenticadas.
  headers:
    CorrelationId:
      description: UUID de correlação aceito ou gerado pela API
      schema:
        type: string
        format: uuid
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    RequiredFrom:
      name: from
      in: query
      required: true
      description: Início inclusivo local, sem offset, interpretado por APP_TIME_ZONE
      schema:
        type: string
        examples: ["2030-01-15T10:00:00"]
    RequiredTo:
      name: to
      in: query
      required: true
      description: Fim exclusivo local, sem offset, interpretado por APP_TIME_ZONE
      schema:
        type: string
        examples: ["2030-01-15T11:00:00"]
...
    AppointmentId:
      name: appointmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BlockId:
      name: blockId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Problem Details seguro; status e tipo dependem da falha
      headers:
        X-Correlation-ID:
          $ref: "#/components/headers/CorrelationId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AppointmentStatus:
      type: string
      enum: [SCHEDULED, CONFIRMED, COMPLETED, CANCELLED, NO_SHOW]
    CreateAppointmentRequest:
      type: object
      required: [patientId, startsAt, durationMinutes]
      additionalProperties: false
      properties:
        patientId:
          type: string
          format: uuid
        startsAt:
          type: string
          description: Data local sem offset
          examples: ["2030-01-15T10:00:00"]
        durationMinutes:
          type: integer
          enum: [15, 30, 45, 60]
    RescheduleAppointmentRequest:
      type: object
      required: [startsAt, durationMinutes, version]
      additionalProperties: false
      properties:
        startsAt:
          type: string
          description: Data local sem offset
          examples: ["2030-01-15T10:00:00"]
        durationMinutes:
          type: integer
          enum: [15, 30, 45, 60]
        version:
          type: integer
          format: int64
          minimum: 0
    ChangeAppointmentStatusRequest:
      type: object
      required: [status, version]
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        version:
          type: integer
          format: int64
          minimum: 0
    Appointment:
      type: object
      required: [id, patientId, startsAt, endsAt, durationMinutes, status, version]
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
          description: Data com offset
        endsAt:
          type: string
          format: date-time
          description: Data com offset
        durationMinutes:
          type: integer
          enum: [15, 30, 45, 60]
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        version:
          type: integer
          format: int64
    AppointmentPage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/Appointment"
    CreateScheduleBlockRequest:
      type: object
      required: [startsAt, endsAt, reason]
      additionalProperties: false
      properties:
        startsAt:
          type: string
          description: Data local sem offset
          examples: ["2030-01-15T10:00:00"]
        endsAt:
          type: string
          description: Data local sem offset
          examples: ["2030-01-15T11:00:00"]
        reason:
          type: string
          minLength: 1
    ScheduleBlock:
      type: object
      required: [id, startsAt, endsAt, reason, createdAt]
      properties:
        id:
          type: string
          format: uuid
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    Problem:
      type: object
      required: [type, title, status, detail, correlationId]
      properties:
        type:
          type: string
          format: uri
          examples: [urn:problem:invalid-request]
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        correlationId:
          type: string
          format: uuid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"